chore: remove old schema types
